+++ pocket_ic/pocket_ic.py
@@ -7,7 +7,7 @@ import base64
 import ic
 from enum import Enum
 from ic.candid import Types
-from typing import List, Optional, Any
+from typing import Optional, Any
 from pocket_ic.pocket_ic_server import PocketICServer
 
 
@@ -21,7 +21,7 @@ class SubnetKind(Enum):
     NNS = "NNS"
     SNS = "SNS"
     SYSTEM = "System"
-
+    VERIFIED_APPLICATION = "VerifiedApplication"  
 
 class SubnetConfig:
     """The configuration of subnets for a PocketIC instance."""
@@ -35,24 +35,25 @@ class SubnetConfig:
         nns=False,
         sns=False,
         system=0,
+        verified_application=0,
     ) -> None:
-        self.application = application
-        self.bitcoin = bitcoin
-        self.fiduciary = fiduciary
-        self.ii = ii
-        self.nns = nns
-        self.sns = sns
-        self.system = system
+        self.application = [{ "state_config": "New", "instruction_config": "Production" }] * application
+        self.bitcoin = { "state_config": "New", "instruction_config": "Production" } if bitcoin else None
+        self.fiduciary = { "state_config": "New", "instruction_config": "Production" } if fiduciary else None
+        self.ii = { "state_config": "New", "instruction_config": "Production" } if ii else None
+        self.nns = { "state_config": "New", "instruction_config": "Production" } if nns else None
+        self.sns = { "state_config": "New", "instruction_config": "Production" } if sns else None
+        self.system = [{ "state_config": "New", "instruction_config": "Production" }] * system
+        self.verified_application = [{ "state_config": "New", "instruction_config": "Production" }] * verified_application
 
     def __repr__(self) -> str:
-        return f"SubnetConfigSet(application={self.application}, bitcoin={self.bitcoin}, fiduciary={self.fiduciary}, ii={self.ii}, nns={self.nns}, sns={self.sns}, system={self.system})"
+        return f"SubnetConfigSet(application={self.application}, bitcoin={self.bitcoin}, fiduciary={self.fiduciary}, ii={self.ii}, nns={self.nns}, sns={self.sns}, system={self.system}, verified_application={self.verified_application})"
 
     def validate(self) -> None:
         """Validates the subnet configuration.
 
         Raises:
-            ValueError: if no subnet is configured or if the number of application or system
-                subnets is negative
+            ValueError: if no subnet is configured
         """
         if not (
             self.bitcoin
@@ -62,34 +63,65 @@ class SubnetConfig:
             or self.sns
             or self.system
             or self.application
+            or self.verified_application
         ):
             raise ValueError("At least one subnet must be configured.")
 
-        if self.application < 0 or self.system < 0:
-            raise ValueError(
-                "The number of application and system subnets must be non-negative."
-            )
+    def add_subnet_with_state(self, subnet_type: SubnetKind, state_dir_path: str, nns_subnet_id: ic.Principal):
+        """Add a subnet with state loaded form the given state directory.
+        Note that the provided path must be accessible for the PocketIC server process.
 
-    def with_nns_state(self, state_dir_path: str, nns_subnet_id: ic.Principal):
-        """Provide an NNS state directory and a subnet id. """
-        self.nns = (state_dir_path, nns_subnet_id)
+        `state_dir` should point to a directory which is expected to have the following structure:
+       
+        state_dir/
+         |-- backups
+         |-- checkpoints
+         |-- diverged_checkpoints
+         |-- diverged_state_markers
+         |-- fs_tmp
+         |-- page_deltas
+         |-- states_metadata.pbuf
+         |-- tip
+         `-- tmp
+       
+        `subnet_id` should be the subnet ID of the subnet in the state to be loaded"""
+
+        raw_subnet_id = base64.b64encode(nns_subnet_id.bytes).decode()
+
+        match subnet_type:
+            case SubnetKind.APPLICATION:
+                self.application.append({"state_config": {"FromPath": [state_dir_path, {"subnet_id": raw_subnet_id}]}, "instruction_config": "Production" })
+            case SubnetKind.BITCOIN:
+                self.bitcoin = {"state_config": {"FromPath": [state_dir_path, {"subnet_id": raw_subnet_id}]}, "instruction_config": "Production" }
+            case SubnetKind.FIDUCIARY:
+                self.fiduciary = {"state_config": {"FromPath": [state_dir_path, {"subnet_id": raw_subnet_id}]}, "instruction_config": "Production" }
+            case SubnetKind.II:
+                self.ii = {"state_config": {"FromPath": [state_dir_path, {"subnet_id": raw_subnet_id}]}, "instruction_config": "Production" }
+            case SubnetKind.NNS:
+                self.nns = {"state_config": {"FromPath": [state_dir_path, {"subnet_id": raw_subnet_id}]}, "instruction_config": "Production" }
+            case SubnetKind.SNS:
+                self.sns = {"state_config": {"FromPath": [state_dir_path, {"subnet_id": raw_subnet_id}]}, "instruction_config": "Production" }
+            case SubnetKind.SYSTEM:
+                self.system.append({"state_config": {"FromPath": [state_dir_path, {"subnet_id": raw_subnet_id}]}, "instruction_config": "Production" })
+            case SubnetKind.VERIFIED_APPLICATION:
+                self.verified_application.append({"state_config": {"FromPath": [state_dir_path, {"subnet_id": raw_subnet_id}]}, "instruction_config": "Production" })
 
     def _json(self) -> dict:
-        if isinstance(self.nns, tuple):
-            raw_subnet_id = base64.b64encode(self.nns[1].bytes).decode()
-            nns = {"FromPath": (self.nns[0], {"subnet_id": raw_subnet_id})}
-        elif self.nns:
-            nns = "New"
-        else:
-            nns = None
         return {
-            "application": self.application * ["New"],
-            "bitcoin": "New" if self.bitcoin else None,
-            "fiduciary": "New" if self.fiduciary else None,
-            "ii": "New" if self.ii else None,
-            "nns": nns,
-            "sns": "New" if self.sns else None,
-            "system": self.system * ["New"],
+            "subnet_config_set": {
+                "application": self.application,
+                "bitcoin": self.bitcoin,
+                "fiduciary": self.fiduciary,
+                "ii": self.ii,
+                "nns": self.nns,
+                "sns": self.sns,
+                "system": self.system,
+                "verified_application": self.verified_application,
+            },
+            "state_dir": None,
+            "nonmainnet_features": False,
+            "log_level": None,
+            "bitcoind_addr": None
         }
 
 
@@ -501,7 +533,8 @@ class PocketIC:
 
     def _generate_topology(self, topology):
         t = dict()
-        for subnet_id, config in topology.items():
+        subnets = topology['subnet_configs']
+        for subnet_id, config in subnets.items():
             subnet_id = ic.Principal.from_str(subnet_id)
             subnet_kind = SubnetKind(config["subnet_kind"])
             t.update({subnet_id: subnet_kind})